refactor: DataAccessException을 상속한 상세 예외 추가
