+++ jdbc/src/main/java/nextstep/jdbc/exception/InvalidSqlException.java
@@ -1,0 +1,8 @@
+package nextstep.jdbc.exception;
+
+public class InvalidSqlException extends DataAccessException {
+
+    public InvalidSqlException(String message) {
+        super(message);
+    }
+}
