+++ jdbc/src/main/java/nextstep/jdbc/exception/NoSuchDataException.java
@@ -1,0 +1,8 @@
+package nextstep.jdbc.exception;
+
+public class NoSuchDataException extends DataAccessException {
+
+    public NoSuchDataException(String message) {
+        super(message);
+    }
+}
